modules/ble/desire_advertiser: add non thread based initiation
Instead of creating an event thread with its own thread queue, pass
the queue on initialization. This allows to use the same event thread
for handling multiple event, saving up RAM. If a specific thread wants
to be used this can still be done by using the
'desire_advertiser_threaded' module.