modules/sys/crypto_manager: revert using hacl
Revert defaulting to using hacl since it required 2*THREAD_STACKSIZE_MAIN
which currently require increasing in the test application every time
a crypto_manager function is called. This is done every time
bootstrap_new_epoch is called. The best would be to offload this to
the event thread in all cases, so at init but also during time
correction (currently this is in the BLE callback).