modules/sys/*_ed: fix *ed_list_finish loop
Currently there was a bug where if the last node in the list was
discarded then the previous one would be re-handled.

This fixes it by checking at the start of the loop if it should be
the last iteration. If a node is removed than it is reset to the
node from the previous iteration making sure all nodes get handled
even if removing the one before last.